Rename nav-link-container for consistency Move nav-bar hide logic to header
